Allowed for output scale specifed
Pixels are now only sampled at a specified output scale (e.g. 1500m) where there are a minimum number of valid MCD15 pixels (500m) and Landsat pixels (30m)
This is done to ensure that if we aggregate MCD15 and LC data we ensure the pixels we sample are almost all unmasked values.